Move the Posting class to it's own file
