+++ app/src/main/java/be/chvp/nanoledger/data/Transaction.kt
@@ -1,15 +1,5 @@
 package be.chvp.nanoledger.data
 
-data class Amount(val quantity: String, val currency: String, val original: String)
-
-data class Posting(
-    val account: String?,
-    val amount: Amount?,
-    val note: String?,
-) {
-    fun contains(query: String) = account?.contains(query, ignoreCase = true) ?: false
-}
-
 data class Transaction(
     val firstLine: Int,
     val lastLine: Int,
